+++ Backend/src/routes/billing.ts
@@ -15,7 +15,7 @@ export default new Elysia({prefix: '/billing'})
     const relativePath = path.slice(8);
     const mustLog = ['', '/', '/portal', '/webhook'];
 
-    if (mustLog.some((p) => relativePath === p) && user) {
+    if (mustLog.some((p) => relativePath === p) && !user) {
       return error(401, 'You are not logged in');
     }
   })
@@ -32,7 +32,7 @@ export default new Elysia({prefix: '/billing'})
       return_url: mode === 'dev' ? 'http://localhost:5173/fail' : 'https://shadowself.io/fail',
       line_items: [
         {
-          price: pricingModal[type].priceID,
+          price: pricingModal[type],
           quantity: 1,
         },
       ],