+++ Frontend/src/routes/purchase/+page.server.ts
@@ -1,10 +1,14 @@
 import type {PageServerLoad, Actions} from './$types';
-import {loadStripe} from '@stripe/stripe-js';
+import {fetchApi} from '$lib';
 
 export const load: PageServerLoad = async () => {
-  return {stripe: await loadStripe(process.env.STRIPE_PUBLISHABLE_KEY!)};
+  return {stripeKey: process.env.STRIPE_PUBLISHABLE_KEY};
 };
 
 export const actions = {
-  default: async ({request}) => {},
+  default: async ({request}) => {
+    const formData = await request.formData();
+    const type = formData.get('type')?.toString().toLowerCase();
+    return await fetchApi(`/billing?type=${type}`, 'GET');
+  },
 } satisfies Actions;