+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.07] - 2024-12-22
+
+### Added
+
+- Add stripe as payment processor
+- Add stripe keys and prices
+- Add billing backend route
+- Add checkout session
+- Add purchase page
+
 ## [0.3.06] - 2024-12-12
 
 ### Added